+++ torch/_functorch/eager_transforms.py
@@ -34,6 +34,7 @@ from torch._C._functorch import (
     get_inplace_requires_grad_allowed
 )
 
+
 argnums_t = Union[int, Tuple[int, ...]]
 
 
@@ -165,100 +166,6 @@ def _autograd_grad(outputs, inputs, grad_outputs=None, retain_graph=False, creat
 
 # How do we increment and decrement the nesting? I don't think we can.
 def vjp(func: Callable, *primals, has_aux: bool = False):
-    """
-    Standing for the vector-Jacobian product, returns a tuple containing the
-    results of :attr:`func` applied to :attr:`primals` and a function that, when
-    given ``cotangents``, computes the reverse-mode Jacobian of :attr:`func` with
-    respect to :attr:`primals` times ``cotangents``.
-
-    Args:
-        func (Callable): A Python function that takes one or more arguments. Must
-            return one or more Tensors.
-        primals (Tensors): Positional arguments to :attr:`func` that must all be
-            Tensors. The returned function will also be computing the
-            derivative with respect to these arguments
-        has_aux (bool): Flag indicating that :attr:`func` returns a
-            ``(output, aux)`` tuple where the first element is the output of
-            the function to be differentiated and the second element is
-            other auxiliary objects that will not be differentiated.
-            Default: False.
-
-    Returns:
-        Returns a ``(output, vjp_fn)`` tuple containing the output of :attr:`func`
-        applied to :attr:`primals` and a function that computes the vjp of
-        :attr:`func` with respect to all :attr:`primals` using the cotangents passed
-        to the returned function. If ``has_aux is True``, then instead returns a
-        ``(output, vjp_fn, aux)`` tuple.
-        The returned ``vjp_fn`` function will return a tuple of each VJP.
-
-    When used in simple cases, :func:`vjp` behaves the same as :func:`grad`
-
-        >>> x = torch.randn([5])
-        >>> f = lambda x: x.sin().sum()
-        >>> (_, vjpfunc) = functorch.vjp(f, x)
-        >>> grad = vjpfunc(torch.tensor(1.))[0]
-        >>> assert torch.allclose(grad, functorch.grad(f)(x))
-
-    However, :func:`vjp` can support functions with multiple outputs by
-    passing in the cotangents for each of the outputs
-
-        >>> x = torch.randn([5])
-        >>> f = lambda x: (x.sin(), x.cos())
-        >>> (_, vjpfunc) = functorch.vjp(f, x)
-        >>> vjps = vjpfunc((torch.ones([5]), torch.ones([5])))
-        >>> assert torch.allclose(vjps[0], x.cos() + -x.sin())
-
-    :func:`vjp` can even support outputs being Python structs
-
-        >>> x = torch.randn([5])
-        >>> f = lambda x: {'first': x.sin(), 'second': x.cos()}
-        >>> (_, vjpfunc) = functorch.vjp(f, x)
-        >>> cotangents = {'first': torch.ones([5]), 'second': torch.ones([5])}
-        >>> vjps = vjpfunc(cotangents)
-        >>> assert torch.allclose(vjps[0], x.cos() + -x.sin())
-
-    The function returned by :func:`vjp` will compute the partials with
-    respect to each of the :attr:`primals`
-
-        >>> x, y = torch.randn([5, 4]), torch.randn([4, 5])
-        >>> (_, vjpfunc) = functorch.vjp(torch.matmul, x, y)
-        >>> cotangents = torch.randn([5, 5])
-        >>> vjps = vjpfunc(cotangents)
-        >>> assert len(vjps) == 2
-        >>> assert torch.allclose(vjps[0], torch.matmul(cotangents, y.transpose(0, 1)))
-        >>> assert torch.allclose(vjps[1], torch.matmul(x.transpose(0, 1), cotangents))
-
-    :attr:`primals` are the positional arguments for :attr:`f`. All kwargs use their
-    default value
-
-        >>> x = torch.randn([5])
-        >>> def f(x, scale=4.):
-        >>>   return x * scale
-        >>>
-        >>> (_, vjpfunc) = functorch.vjp(f, x)
-        >>> vjps = vjpfunc(torch.ones_like(x))
-        >>> assert torch.allclose(vjps[0], torch.full(x.shape, 4.))
-
-    .. note::
-        Using PyTorch ``torch.no_grad`` together with ``vjp``.
-        Case 1: Using ``torch.no_grad`` inside a function:
-
-            >>> def f(x):
-            >>>     with torch.no_grad():
-            >>>         c = x ** 2
-            >>>     return x - c
-
-        In this case, ``vjp(f)(x)`` will respect the inner ``torch.no_grad``.
-
-        Case 2: Using ``vjp`` inside ``torch.no_grad`` context manager:
-
-            >>> with torch.no_grad():
-            >>>     vjp(f)(x)
-
-        In this case, ``vjp`` will respect the inner ``torch.no_grad``, but not the
-        outer one. This is because ``vjp`` is a "function transform": its result
-        should not depend on the result of a context manager outside of ``f``.
-    """
     return _vjp_with_argnums(func, *primals, has_aux=has_aux)
 
 
@@ -339,121 +246,6 @@ def _safe_zero_index(x):
 
 
 def jacrev(func: Callable, argnums: Union[int, Tuple[int]] = 0, *, has_aux=False):
-    """
-    Computes the Jacobian of :attr:`func` with respect to the arg(s) at index
-    :attr:`argnum` using reverse mode autodiff
-
-    Args:
-        func (function): A Python function that takes one or more arguments,
-            one of which must be a Tensor, and returns one or more Tensors
-        argnums (int or Tuple[int]): Optional, integer or tuple of integers,
-            saying which arguments to get the Jacobian with respect to.
-            Default: 0.
-        has_aux (bool): Flag indicating that :attr:`func` returns a
-            ``(output, aux)`` tuple where the first element is the output of
-            the function to be differentiated and the second element is
-            auxiliary objects that will not be differentiated.
-            Default: False.
-
-    Returns:
-        Returns a function that takes in the same inputs as :attr:`func` and
-        returns the Jacobian of :attr:`func` with respect to the arg(s) at
-        :attr:`argnums`. If ``has_aux is True``, then the returned function
-        instead returns a ``(jacobian, aux)`` tuple where ``jacobian``
-        is the Jacobian and ``aux`` is auxiliary objects returned by :attr:`func`.
-
-    A basic usage with a pointwise, unary operation will give a diagonal array
-    as the Jacobian
-
-        >>> from functorch import jacrev
-        >>> x = torch.randn(5)
-        >>> jacobian = jacrev(torch.sin)(x)
-        >>> expected = torch.diag(torch.cos(x))
-        >>> assert torch.allclose(jacobian, expected)
-
-    If you would like to compute the output of the function as well as the
-    jacobian of the function, use the ``has_aux`` flag to return the output
-    as an auxiliary object:
-
-        >>> from functorch import jacrev
-        >>> x = torch.randn(5)
-        >>>
-        >>> def f(x):
-        >>>   return x.sin()
-        >>>
-        >>> def g(x):
-        >>>   result = f(x)
-        >>>   return result, result
-        >>>
-        >>> jacobian_f, f_x = jacrev(g, has_aux=True)(x)
-        >>> assert torch.allclose(f_x, f(x))
-
-    :func:`jacrev` can be composed with vmap to produce batched
-    Jacobians:
-
-        >>> from functorch import jacrev, vmap
-        >>> x = torch.randn(64, 5)
-        >>> jacobian = vmap(jacrev(torch.sin))(x)
-        >>> assert jacobian.shape == (64, 5, 5)
-
-    Additionally, :func:`jacrev` can be composed with itself to produce
-    Hessians
-
-        >>> from functorch import jacrev
-        >>> def f(x):
-        >>>   return x.sin().sum()
-        >>>
-        >>> x = torch.randn(5)
-        >>> hessian = jacrev(jacrev(f))(x)
-        >>> assert torch.allclose(hessian, torch.diag(-x.sin()))
-
-    By default, :func:`jacrev` computes the Jacobian with respect to the first
-    input. However, it can compute the Jacboian with respect to a different
-    argument by using :attr:`argnums`:
-
-        >>> from functorch import jacrev
-        >>> def f(x, y):
-        >>>   return x + y ** 2
-        >>>
-        >>> x, y = torch.randn(5), torch.randn(5)
-        >>> jacobian = jacrev(f, argnums=1)(x, y)
-        >>> expected = torch.diag(2 * y)
-        >>> assert torch.allclose(jacobian, expected)
-
-    Additionally, passing a tuple to :attr:`argnums` will compute the Jacobian
-    with respect to multiple arguments
-
-        >>> from functorch import jacrev
-        >>> def f(x, y):
-        >>>   return x + y ** 2
-        >>>
-        >>> x, y = torch.randn(5), torch.randn(5)
-        >>> jacobian = jacrev(f, argnums=(0, 1))(x, y)
-        >>> expectedX = torch.diag(torch.ones_like(x))
-        >>> expectedY = torch.diag(2 * y)
-        >>> assert torch.allclose(jacobian[0], expectedX)
-        >>> assert torch.allclose(jacobian[1], expectedY)
-
-    .. note::
-        Using PyTorch ``torch.no_grad`` together with ``jacrev``.
-        Case 1: Using ``torch.no_grad`` inside a function:
-
-            >>> def f(x):
-            >>>     with torch.no_grad():
-            >>>         c = x ** 2
-            >>>     return x - c
-
-        In this case, ``jacrev(f)(x)`` will respect the inner ``torch.no_grad``.
-
-        Case 2: Using ``jacrev`` inside ``torch.no_grad`` context manager:
-
-            >>> with torch.no_grad():
-            >>>     jacrev(f)(x)
-
-        In this case, ``jacrev`` will respect the inner ``torch.no_grad``, but not the
-        outer one. This is because ``jacrev`` is a "function transform": its result
-        should not depend on the result of a context manager outside of ``f``.
-    """
     @wraps(func)
     def wrapper_fn(*args):
         vjp_out = _vjp_with_argnums(func, *args, argnums=argnums, has_aux=has_aux)
@@ -735,56 +527,6 @@ def safe_unpack_dual(dual, strict):
 
 
 def jvp(func: Callable, primals: Any, tangents: Any, *, strict: bool = False, has_aux: bool = False):
-    """
-    Standing for the Jacobian-vector product, returns a tuple containing
-    the output of `func(*primals)` and the "Jacobian of ``func`` evaluated at
-    ``primals``" times ``tangents``. This is also known as forward-mode autodiff.
-
-    Args:
-        func (function): A Python function that takes one or more arguments,
-            one of which must be a Tensor, and returns one or more Tensors
-        primals (Tensors): Positional arguments to :attr:`func` that must all be
-            Tensors. The returned function will also be computing the
-            derivative with respect to these arguments
-        tangents (Tensors): The "vector" for which Jacobian-vector-product is
-            computed. Must be the same structure and sizes as the inputs to
-            ``func``.
-        has_aux (bool): Flag indicating that :attr:`func` returns a
-            ``(output, aux)`` tuple where the first element is the output of
-            the function to be differentiated and the second element is
-            other auxiliary objects that will not be differentiated.
-            Default: False.
-
-    Returns:
-        Returns a ``(output, jvp_out)`` tuple containing the output of ``func``
-        evaluated at ``primals`` and the Jacobian-vector product.
-        If ``has_aux is True``, then instead returns a ``(output, jvp_out, aux)`` tuple.
-
-    .. note::
-        You may see this API error out with "forward-mode AD not implemented
-        for operator X". If so, please file a bug report and we will prioritize it.
-
-    jvp is useful when you wish to compute gradients of a function R^1 -> R^N
-
-        >>> from functorch import jvp
-        >>> x = torch.randn([])
-        >>> f = lambda x: x * torch.tensor([1., 2., 3])
-        >>> value, grad = jvp(f, (x,), (torch.tensor(1.),))
-        >>> assert torch.allclose(value, f(x))
-        >>> assert torch.allclose(grad, torch.tensor([1., 2, 3]))
-
-    :func:`jvp` can support functions with multiple inputs by passing in the
-    tangents for each of the inputs
-
-         >>> from functorch import jvp
-         >>> x = torch.randn(5)
-         >>> y = torch.randn(5)
-         >>> f = lambda x, y: (x * y)
-         >>> _, output = jvp(f, (x, y), (torch.ones(5), torch.ones(5)))
-         >>> assert torch.allclose(output, x + y)
-
-    """
-
     return _jvp_with_argnums(func, primals, tangents, argnums=None, strict=strict, has_aux=has_aux)
 
 
@@ -873,110 +615,6 @@ def safe_unflatten(tensor, dim, shape):
 
 
 def jacfwd(func: Callable, argnums: argnums_t = 0, has_aux: bool = False, *, randomness: str = "error"):
-    """
-    Computes the Jacobian of :attr:`func` with respect to the arg(s) at index
-    :attr:`argnum` using forward-mode autodiff
-
-    Args:
-        func (function): A Python function that takes one or more arguments,
-            one of which must be a Tensor, and returns one or more Tensors
-        argnums (int or Tuple[int]): Optional, integer or tuple of integers,
-            saying which arguments to get the Jacobian with respect to.
-            Default: 0.
-        has_aux (bool): Flag indicating that :attr:`func` returns a
-            ``(output, aux)`` tuple where the first element is the output of
-            the function to be differentiated and the second element is
-            auxiliary objects that will not be differentiated.
-            Default: False.
-        randomness(str): Flag indicating what type of randomness to use.
-            See :func:`vmap` for more detail. Allowed: "different", "same", "error".
-            Default: "error"
-
-    Returns:
-        Returns a function that takes in the same inputs as :attr:`func` and
-        returns the Jacobian of :attr:`func` with respect to the arg(s) at
-        :attr:`argnums`. If ``has_aux is True``, then the returned function
-        instead returns a ``(jacobian, aux)`` tuple where ``jacobian``
-        is the Jacobian and ``aux`` is auxiliary objects returned by :attr:`func`.
-
-    .. note::
-        You may see this API error out with "forward-mode AD not implemented
-        for operator X". If so, please file a bug report and we will prioritize it.
-        An alternative is to use :func:`jacrev`, which has better operator coverage.
-
-    A basic usage with a pointwise, unary operation will give a diagonal array
-    as the Jacobian
-
-        >>> from functorch import jacfwd
-        >>> x = torch.randn(5)
-        >>> jacobian = jacfwd(torch.sin)(x)
-        >>> expected = torch.diag(torch.cos(x))
-        >>> assert torch.allclose(jacobian, expected)
-
-    :func:`jacfwd` can be composed with vmap to produce batched
-    Jacobians:
-
-        >>> from functorch import jacfwd, vmap
-        >>> x = torch.randn(64, 5)
-        >>> jacobian = vmap(jacfwd(torch.sin))(x)
-        >>> assert jacobian.shape == (64, 5, 5)
-
-    If you would like to compute the output of the function as well as the
-    jacobian of the function, use the ``has_aux`` flag to return the output
-    as an auxiliary object:
-
-        >>> from functorch import jacfwd
-        >>> x = torch.randn(5)
-        >>>
-        >>> def f(x):
-        >>>   return x.sin()
-        >>>
-        >>> def g(x):
-        >>>   result = f(x)
-        >>>   return result, result
-        >>>
-        >>> jacobian_f, f_x = jacfwd(g, has_aux=True)(x)
-        >>> assert torch.allclose(f_x, f(x))
-
-    Additionally, :func:`jacrev` can be composed with itself or :func:`jacrev`
-    to produce Hessians
-
-        >>> from functorch import jacfwd, jacrev
-        >>> def f(x):
-        >>>   return x.sin().sum()
-        >>>
-        >>> x = torch.randn(5)
-        >>> hessian = jacfwd(jacrev(f))(x)
-        >>> assert torch.allclose(hessian, torch.diag(-x.sin()))
-
-    By default, :func:`jacfwd` computes the Jacobian with respect to the first
-    input. However, it can compute the Jacboian with respect to a different
-    argument by using :attr:`argnums`:
-
-        >>> from functorch import jacfwd
-        >>> def f(x, y):
-        >>>   return x + y ** 2
-        >>>
-        >>> x, y = torch.randn(5), torch.randn(5)
-        >>> jacobian = jacfwd(f, argnums=1)(x, y)
-        >>> expected = torch.diag(2 * y)
-        >>> assert torch.allclose(jacobian, expected)
-
-    Additionally, passing a tuple to :attr:`argnums` will compute the Jacobian
-    with respect to multiple arguments
-
-        >>> from functorch import jacfwd
-        >>> def f(x, y):
-        >>>   return x + y ** 2
-        >>>
-        >>> x, y = torch.randn(5), torch.randn(5)
-        >>> jacobian = jacfwd(f, argnums=(0, 1))(x, y)
-        >>> expectedX = torch.diag(torch.ones_like(x))
-        >>> expectedY = torch.diag(2 * y)
-        >>> assert torch.allclose(jacobian[0], expectedX)
-        >>> assert torch.allclose(jacobian[1], expectedY)
-
-    """
     @wraps(func)
     def wrapper_fn(*args):
         primals = args if argnums is None else _slice_argnums(args, argnums)
@@ -1025,76 +663,7 @@ def jacfwd(func: Callable, argnums: argnums_t = 0, has_aux: bool = False, *, ran
     return wrapper_fn
 
 
-def hessian(func, argnums=0):
-    """
-    Computes the Hessian of :attr:`func` with respect to the arg(s) at index
-    :attr:`argnum` via a forward-over-reverse strategy.
-
-    The forward-over-reverse strategy (composing ``jacfwd(jacrev(func))``) is
-    a good default for good performance. It is possible to compute Hessians
-    through other compositions of :func:`jacfwd` and :func:`jacrev` like
-    ``jacfwd(jacfwd(func))`` or ``jacrev(jacrev(func))``.
-
-    Args:
-        func (function): A Python function that takes one or more arguments,
-            one of which must be a Tensor, and returns one or more Tensors
-        argnums (int or Tuple[int]): Optional, integer or tuple of integers,
-            saying which arguments to get the Hessian with respect to.
-            Default: 0.
-
-    Returns:
-        Returns a function that takes in the same inputs as :attr:`func` and
-        returns the Hessian of :attr:`func` with respect to the arg(s) at
-        :attr:`argnums`.
-
-    .. note::
-        You may see this API error out with "forward-mode AD not implemented
-        for operator X". If so, please file a bug report and we will prioritize it.
-        An alternative is to use ``jacrev(jacrev(func))``, which has better
-        operator coverage.
-
-    A basic usage with a R^N -> R^1 function gives a N x N Hessian:
-
-        >>> from functorch import hessian
-        >>> def f(x):
-        >>>   return x.sin().sum()
-        >>>
-        >>> x = torch.randn(5)
-        >>> hess = hessian(f)(x)  # equivalent to jacfwd(jacrev(f))(x)
-        >>> assert torch.allclose(hess, torch.diag(-x.sin()))
-
-    """
-    return jacfwd(jacrev(func, argnums), argnums)
-
-
 def grad_and_value(func: Callable, argnums: argnums_t = 0, has_aux: bool = False) -> Callable:
-    """
-    Returns a function to compute a tuple of the gradient and primal, or
-    forward, computation.
-
-    Args:
-        func (Callable): A Python function that takes one or more arguments.
-            Must return a single-element Tensor. If specified :attr:`has_aux`
-            equals ``True``, function can return a tuple of single-element
-            Tensor and other auxiliary objects: ``(output, aux)``.
-        argnums (int or Tuple[int]): Specifies arguments to compute gradients
-            with respect to. :attr:`argnums` can be single integer or tuple of
-            integers. Default: 0.
-        has_aux (bool): Flag indicating that :attr:`func` returns a tensor and
-            other auxiliary objects: ``(output, aux)``. Default: False.
-
-    Returns:
-        Function to compute a tuple of gradients with respect to its inputs
-        and the forward computation. By default, the output of the function is
-        a tuple of the gradient tensor(s) with respect to the first argument
-        and the primal computation. If specified :attr:`has_aux` equals
-        ``True``, tuple of gradients and tuple of the forward computation with
-        output auxiliary objects is returned. If :attr:`argnums` is a tuple of
-        integers, a tuple of a tuple of the output gradients with respect to
-        each :attr:`argnums` value and the forward computation is returned.
-
-    See :func:`grad` for examples
-    """
     @doesnt_support_saved_tensors_hooks
     @wraps(func)
     def wrapper(*args, **kwargs):
@@ -1147,95 +716,6 @@ def grad_and_value(func: Callable, argnums: argnums_t = 0, has_aux: bool = False
 
 
 def grad(func: Callable, argnums: argnums_t = 0, has_aux: bool = False) -> Callable:
-    """``grad`` operator helps computing gradients of :attr:`func` with respect to the
-    input(s) specified by :attr:`argnums`. This operator can be nested to
-    compute higher-order gradients.
-
-    Args:
-        func (Callable): A Python function that takes one or more arguments.
-            Must return a single-element Tensor. If specified :attr:`has_aux` equals ``True``,
-            function can return a tuple of single-element Tensor and other auxiliary objects:
-            ``(output, aux)``.
-        argnums (int or Tuple[int]): Specifies arguments to compute gradients with respect to.
-            :attr:`argnums` can be single integer or tuple of integers. Default: 0.
-        has_aux (bool): Flag indicating that :attr:`func` returns a tensor and other
-            auxiliary objects: ``(output, aux)``. Default: False.
-
-    Returns:
-        Function to compute gradients with respect to its inputs. By default, the output of
-        the function is the gradient tensor(s) with respect to the first argument.
-        If specified :attr:`has_aux` equals ``True``, tuple of gradients and output auxiliary objects
-        is returned. If :attr:`argnums` is a tuple of integers, a tuple of output gradients with
-        respect to each :attr:`argnums` value is returned.
-
-    Example of using ``grad``:
-
-        >>> from functorch import grad
-        >>> x = torch.randn([])
-        >>> cos_x = grad(lambda x: torch.sin(x))(x)
-        >>> assert torch.allclose(cos_x, x.cos())
-        >>>
-        >>> # Second-order gradients
-        >>> neg_sin_x = grad(grad(lambda x: torch.sin(x)))(x)
-        >>> assert torch.allclose(neg_sin_x, -x.sin())
-
-    When composed with ``vmap``, ``grad`` can be used to compute per-sample-gradients:
-
-        >>> from functorch import grad
-        >>> from functorch import vmap
-        >>> batch_size, feature_size = 3, 5
-        >>>
-        >>> def model(weights, feature_vec):
-        >>>     # Very simple linear model with activation
-        >>>     assert feature_vec.dim() == 1
-        >>>     return feature_vec.dot(weights).relu()
-        >>>
-        >>> def compute_loss(weights, example, target):
-        >>>     y = model(weights, example)
-        >>>     return ((y - target) ** 2).mean()  # MSELoss
-        >>>
-        >>> weights = torch.randn(feature_size, requires_grad=True)
-        >>> examples = torch.randn(batch_size, feature_size)
-        >>> targets = torch.randn(batch_size)
-        >>> inputs = (weights, examples, targets)
-        >>> grad_weight_per_example = vmap(grad(compute_loss), in_dims=(None, 0, 0))(*inputs)
-
-    Example of using ``grad`` with :attr:`has_aux` and :attr:`argnums`:
-
-        >>> from functorch import grad
-        >>> def my_loss_func(y, y_pred):
-        >>>    loss_per_sample = (0.5 * y_pred - y) ** 2
-        >>>    loss = loss_per_sample.mean()
-        >>>    return loss, (y_pred, loss_per_sample)
-        >>>
-        >>> fn = grad(my_loss_func, argnums=(0, 1), has_aux=True)
-        >>> y_true = torch.rand(4)
-        >>> y_preds = torch.rand(4, requires_grad=True)
-        >>> out = fn(y_true, y_preds)
-        >>> > output is ((grads w.r.t y_true, grads w.r.t y_preds), (y_pred, loss_per_sample))
-
-    .. note::
-        Using PyTorch ``torch.no_grad`` together with ``grad``.
-
-        Case 1: Using ``torch.no_grad`` inside a function:
-
-            >>> def f(x):
-            >>>     with torch.no_grad():
-            >>>         c = x ** 2
-            >>>     return x - c
-
-        In this case, ``grad(f)(x)`` will respect the inner ``torch.no_grad``.
-
-        Case 2: Using ``grad`` inside ``torch.no_grad`` context manager:
-
-            >>> with torch.no_grad():
-            >>>     grad(f)(x)
-
-        In this case, ``grad`` will respect the inner ``torch.no_grad``, but not the
-        outer one. This is because ``grad`` is a "function transform": its result
-        should not depend on the result of a context manager outside of ``f``.
-
-    """
     @wraps(func)
     def wrapper(*args, **kwargs):
         results = grad_and_value(func, argnums, has_aux=has_aux)(*args, **kwargs)
@@ -1275,158 +755,6 @@ def _unwrap_all_tensors_from_functional(tensor_pytree, *, reapply_views: bool):
 
 
 def functionalize(func: Callable, *, remove: str = 'mutations') -> Callable:
-    """
-    functionalize is a transform that can be used to remove (intermediate)
-    mutations and aliasing from a function, while preserving the function's
-    semantics.
-
-    ``functionalize(func)`` returns a new function with the same semantics
-    as ``func``, but with all intermediate mutations removed.
-    Every inplace operation performed on an intermediate tensor:
-    ``intermediate.foo_()``
-    gets replaced by its out-of-place equivalent:
-    ``intermediate_updated = intermediate.foo()``.
-
-    functionalize is useful for shipping a pytorch program off to
-    backends or compilers that aren't able to easily represent
-    mutations or aliasing operators.
-
-    Args:
-        func (Callable): A Python function that takes one or more arguments.
-        remove (str): An optional string argument, that takes on either
-            the value 'mutations' or 'mutations_and_views'.
-            If 'mutations' is passed in then all mutating operators
-            will be replaced with their non-mutating equivalents.
-            If 'mutations_and_views' is passed in, then additionally, all aliasing
-            operators will be replaced with their non-aliasing equivalents.
-            Default: 'mutations'.
-
-    Returns:
-        Returns a new "functionalized" function. It takes the same inputs as
-        :attr:`func`, and has the same behavior, but any mutations
-        (and optionally aliasing) performed on intermeidate tensors
-        in the function will be removed.
-
-    functionalize will also remove mutations (and views) that were performed on function inputs.
-    However to preserve semantics, functionalize will "fix up" the mutations after
-    the transform has finished running, by detecting if any tensor inputs "should have"
-    been mutated, and copying the new data back to the inputs if necessary.
-
-
-    Example::
-
-        >>> import torch
-        >>> from functorch import make_fx
-        >>> from functorch.experimental import functionalize
-        >>>
-        >>> A function that uses mutations and views, but only on intermediate tensors.
-        >>> def f(a):
-        ...     b = a + 1
-        ...     c = b.view(-1)
-        ...     c.add_(1)
-        ...     return b
-        ...
-        >>> inpt = torch.randn(2)
-        >>>
-        >>> out1 = f(inpt)
-        >>> out2 = functionalize(f)(inpt)
-        >>>
-        >>> # semantics are the same (outputs are equivalent)
-        >>> print(torch.allclose(out1, out2))
-        True
-        >>>
-        >>> f_traced = make_fx(f)(inpt)
-        >>> f_no_mutations_traced = make_fx(functionalize(f))(inpt)
-        >>> f_no_mutations_and_views_traced = make_fx(functionalize(f, remove='mutations_and_views'))(inpt)
-        >>>
-        >>> print(f_traced.code)
-
-
-
-        def forward(self, a_1):
-            add = torch.ops.aten.add(a_1, 1);  a_1 = None
-            view = torch.ops.aten.view(add, [-1])
-            add_ = torch.ops.aten.add_(view, 1);  view = None
-            return add
-
-        >>> print(f_no_mutations_traced.code)
-
-
-
-        def forward(self, a_1):
-            add = torch.ops.aten.add(a_1, 1);  a_1 = None
-            view = torch.ops.aten.view(add, [-1]);  add = None
-            add_1 = torch.ops.aten.add(view, 1);  view = None
-            view_1 = torch.ops.aten.view(add_1, [2]);  add_1 = None
-            return view_1
-
-        >>> print(f_no_mutations_and_views_traced.code)
-
-
-
-        def forward(self, a_1):
-            add = torch.ops.aten.add(a_1, 1);  a_1 = None
-            view_copy = torch.ops.aten.view_copy(add, [-1]);  add = None
-            add_1 = torch.ops.aten.add(view_copy, 1);  view_copy = None
-            view_copy_1 = torch.ops.aten.view_copy(add_1, [2]);  add_1 = None
-            return view_copy_1
-
-
-        >>> A function that mutates its input tensor
-        >>> def f(a):
-        ...     b = a.view(-1)
-        ...     b.add_(1)
-        ...     return a
-        ...
-        >>> f_no_mutations_and_views_traced = make_fx(functionalize(f, remove='mutations_and_views'))(inpt)
-        >>>
-        >>> All mutations and views have been removed,
-        >>> but there is an extra copy_ in the graph to correctly apply the mutation to the input
-        >>> after the function has completed.
-        >>> print(f_no_mutations_and_views_traced.code)
-
-
-
-        def forward(self, a_1):
-            view_copy = torch.ops.aten.view_copy(a_1, [-1])
-            add = torch.ops.aten.add(view_copy, 1);  view_copy = None
-            view_copy_1 = torch.ops.aten.view_copy(add, [2]);  add = None
-            copy_ = torch.ops.aten.copy_(a_1, view_copy_1);  a_1 = None
-            return view_copy_1
-
-
-    There are a few "failure modes" for functionalize that are worth calling out:
-      (1) Like other functorch transforms, `functionalize()` doesn't work with functions
-          that directly use `.backward()`. The same is true for torch.autograd.grad.
-          If you want to use autograd, you can compute gradients directly
-          with `functionalize(grad(f))`.
-      (2) Like other functorch transforms, `functionalize()` doesn't work with global state.
-          If you call `functionalize(f)` on a function that takes views / mutations of
-          non-local state, functionalization will simply no-op and pass the view/mutation
-          calls directly to the backend.
-          One way to work around this is is to ensure that any non-local state creation
-          is wrapped into a larger function, which you then call functionalize on.
-      (3) `resize_()` has some limitations: functionalize will only work on programs
-          that use resize_()` as long as the tensor being resized is not a view.
-      (4) `as_strided()` has some limitations: functionalize will not work on
-          `as_strided()` calls that result in tensors with overlapping memory.
-
-
-    Finally, a helpful mental model for understanding functionalization is that
-    most user pytorch programs are writting with the public torch API.
-    When executed, torch operators are generally decomposed into
-    our internal C++ "ATen" API.
-    The logic for functionalization happens entirely at the level of ATen.
-    Functionalization knows how to take every aliasing operator in ATen,
-    and map it to its non-aliasing equivalent
-    (e.g. ``tensor.view({-1})`` -> ``at::view_copy(tensor, {-1})``),
-    and how to take every mutating operator in ATen,
-    and map it to its non-mutating equivalent
-    (e.g. ``tensor.add_(1)`` -> ``at::add(tensor, -1)``),
-    while tracking aliases and mutations out-of-line to know when to fix things up.
-    Information about which ATen operators are aliasing or mutating all comes from
-    https://github.com/pytorch/pytorch/blob/master/aten/src/ATen/native/native_functions.yaml.
-    """
     if remove == 'mutations':
         reapply_views = True
     elif remove == 'mutations_and_views':
